# fix: ingest csaf documents in a single transaction